+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Added
 - New sort options to use character code for string comparison
 
+### Changed
+- Renamed to "Stable Sort"
+
 ## [0.2.1-beta] - 2018-09-18
 ### Changed
 - Now uses user environment language as the locale for comparing strings